% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webrSub.R
\name{allsubstr}
\alias{allsubstr}
\title{Return All substring}
\usage{
allsubstr(x, max = NULL)
}
\arguments{
\item{x}{A character string}

\item{max}{maximal number of character allowed}
}
\description{
Return All substring
}
