% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webrSub.R
\name{make_numeric}
\alias{make_numeric}
\title{Delete all non-numeric character}
\usage{
make_numeric(x)
}
\arguments{
\item{x}{A character string to be converted}
}
\description{
Delete all non-numeric character
}
