% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webrSub.R
\name{md2html}
\alias{md2html}
\title{Convert markdown string into html format}
\usage{
md2html(string)
}
\arguments{
\item{string}{The character string to convert}
}
\description{
Convert markdown string into html format
}
