% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webrSub.R
\name{myimport}
\alias{myimport}
\title{Import data file}
\usage{
myimport(file, ...)
}
\arguments{
\item{file}{File to read}

\item{...}{Further argument to be passed to read.csv or import}
}
\description{
Import data file
}
