% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webrSub.R
\name{myLCS_sub}
\alias{myLCS_sub}
\title{Return Longest common string bwteen two strings}
\usage{
myLCS_sub(a, b)
}
\arguments{
\item{a}{A character string}

\item{b}{A character string}
}
\description{
Return Longest common string bwteen two strings
}
