% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webrSub.R
\name{mypaste}
\alias{mypaste}
\title{Paste string with comma}
\usage{
mypaste(tempA, ..., sep = ",")
}
\arguments{
\item{tempA}{a string}

\item{...}{Further argument to be passed to paste()}

\item{sep}{separator}
}
\description{
Paste string with comma
}
