% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webrSub.R
\name{text2vector}
\alias{text2vector}
\title{Make character vector from text}
\usage{
text2vector(text)
}
\arguments{
\item{text}{A character string to be converted}
}
\description{
Make character vector from text
}
