% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webrSub.R
\name{textInput2}
\alias{textInput2}
\title{textInput with disply:inline block}
\usage{
textInput2(inputId, label, value = "", width = 100, ...)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value.}

\item{width}{The width of the input in pixel}

\item{...}{Further argument to be passed to textInput}
}
\description{
textInput with disply:inline block
}
