% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webrSub.R
\name{vector2form}
\alias{vector2form}
\title{Convert string vector to formula}
\usage{
vector2form(vars)
}
\arguments{
\item{vars}{The character string to be converted}
}
\description{
Convert string vector to formula
}
