% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webrSub.R
\name{vector2str}
\alias{vector2str}
\title{Convert string vector to string}
\usage{
vector2str(vars, del = FALSE)
}
\arguments{
\item{vars}{The character string to be converted}

\item{del}{Logical}
}
\description{
Convert string vector to string
}
